requirements.txt
requests==2.31.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
pandas==2.2.2
pytz==2024.1
